Renaming sorted-set file
